fixed and refactored blog details view. Also added links for authors
